Apply forced outage rates from NERC GADS system and reduce PRM
This is more realistic than ignoring forced outages, and tightens up the
capacity supply quite a bit. With this change, we also reduce the PRM,
since forced outages (and spinning reserves and line losses) are included
in the things the NERC PRM is mean to cover. We previously raised the PRM
to the highest level that could be met in 2024, but it is now impossible to
meet any PRM nationwide in the 2024 model (due to lack of resources in zone
p46 in western Wisconsin during one timepoint). So instead we apply a 2%
PRM for future-build (on top of our explicit modeling of forced outages,
spinning reserves and line losses) and need to allow unserved load (with a
penalty) in 2024. Driving extra capacity into the model via forced outages
gives a more accurate representation of how often it will be run than using
a simple PRM. (PRM drives in capacity that will never be used, but forced
outages say 5-10% of capacity will be unavailable all the time, so the
capacity that replaces it will run all the time or nearly so.)