Fix bugs in multi-period model setup
This commit make several changes to make multi-period models work
correctly:

- refactor generator and fuel code so that the GeneratorClusters
  object is created from scratch for each model year. This prevents
  a problem where side-effects in one model year would alter the
  fuel list for a previously run model year.
- provide unique IDs for every timepoint, avoiding a problem introduced
  in the PG-ReEDS refactoring where the same sequential ID would be used
  in two different study periods. All timepoints now have an ID like
  YYYYRMMDDHH, where YYYY = study year, R = repetition of historical
  weather year (e.g., if there are 7 historical weather years, this will
  be a number from 1 to 7 indicating which year the timepoint was sampled
  from), MM DD = calendar month and day the timepoint came from, and
  HH = hour of day for this timepoint (0 is the first hour, beginning
  at midnight)
- add some case definition examples for 2029 and 2035 so it's possible
  to try `python pg_to_switch.py --case s4x1 --year 2029 --year 2030`.

Note that it should theoretically be possible to run multi-period
models with historical and future years (e.g., 2024 and 2030), but then
you will probably need to turn off a lot of the policy rules, since
they currently can't be met in the 2024 model. Future work will add
settings to pg/settings to relax these automatically in historical years,
to avoid this problem.